refactor: Implement co-located test structure for better maintainability
- Move all unit test files to be co-located with their source files
- Update vitest.config.js to find tests in src/ directory
- Fix import paths in all test files to use relative imports
- Keep integration tests and setup files in test/ directory
- Update coverage configuration to exclude config/ directory

New structure:
- src/main.js + src/main.test.js
- src/game/game.js + src/game/game.test.js
- src/entities/bullet.js + src/entities/bullet.test.js
- src/entities/player.js + src/entities/player.test.js
- src/entities/enemies/enemy.js + src/entities/enemies/enemy.test.js
- src/rendering/background.js + src/rendering/background.test.js
- src/rendering/effects.js + src/rendering/effects.test.js
- src/systems/audio.js + src/systems/audio.test.js
- src/systems/powerups.js + src/systems/powerups.test.js
- src/ui/options.js + src/ui/options.test.js
- src/utils/collision.js + src/utils/collision.test.js
- src/utils/math.js + src/utils/math.test.js
- src/constants/game-constants.js + src/constants/game-constants.test.js

Benefits:
- Better discoverability: Tests are right next to source files
- Easier maintenance: Clear relationship between source and test
- Improved developer experience: Faster navigation between files
- Industry standard: Follows modern frontend project conventions

All 330 tests passing, coverage maintained at 78.85%
