+++ src/entities/enemies/enemy.test.js
@@ -4,7 +4,7 @@
  */
 
 import { describe, it, expect, beforeEach, vi } from 'vitest'
-import Enemy from '../src/entities/enemies/enemy.js'
+import Enemy from './enemy.js'
 
 describe('Enemy', () => {
   let mockGame
